+++ R/dicom_reader.R
@@ -72,7 +72,7 @@ dicom_reader <- function(input, tags = list(sop_class_uid = "0008,0016"),
     # MRS is a special case where we need to check for zero padding
     # 5600,0020 - standard dicom MRS data tag
     # 2005,1270 - Philips private dicom MRS data tag
-    if (tag_str %in% c("5600,0020", "2005,1270")) {
+    if (tag_str %in% c("5600,0020", "2005,1270", "0065,FF07", "0065,FF08")) {
       # check for zero padding
       zp <- is.na(pos[7]) & is.na(pos[8])
       if ((vr %in% c(long_vrs, short_vrs)) & zp) {
@@ -133,7 +133,7 @@ dicom_reader <- function(input, tags = list(sop_class_uid = "0008,0016"),
   
     if (debug) {
       tag <- paste("(", tag_str, ")", sep = "")
-      x <- data.frame(tag, start_byte, end_byte, length)
+      x <- data.frame(tag, start_byte, end_byte, length, vr)
       debug_table <- rbind(debug_table, x)
     }