+++ R/mrs_read_dicom.R
@@ -31,10 +31,97 @@ read_dicom <- function(fname, verbose, extra) {
       stop(paste0("Unsupported SOP class UID : ", sop_class_uid,
                   ". This doesn't look like MRS data."))
     }
+  } else if (grepl("UIH", manuf)) {
+      return(read_uih_dicom(fraw, extra, verbose))
   } else {
     stop(paste0("DICOM MRS manufacturer format: \"", trimws(manuf),
                 "\" is not currently supported."))
   }
+}
+
+read_uih_dicom <- function(fraw, extra, verbose) {
+  
+  # list of tags to pull from the dicom file
+  tags <- list(data    = "5600,0020",
+               fs      = "0018,9052",
+               ft      = "0018,9098",
+               ipp     = "0020,0032",
+               iop     = "0020,0037",
+               pixsp   = "0028,0030",
+               slice_t = "0018,0050",
+               rows    = "0028,0010",
+               cols    = "0028,0011",
+               slices  = "0018,9159",
+               Npt     = "0018,9127",
+               tr      = "0018,0080",
+               te      = "0018,0081",
+               seq_n   = "0018,0024")
+
+  # read em out
+  dcm_res <- dicom_reader(fraw, tags)
+  
+  # sort out the right datatypes
+  fs  <- readBin(dcm_res$fs, "double")
+  ft  <- readBin(dcm_res$ft, "double") * 1e6
+  ipp <- as.numeric(strsplit(rawToChar(dcm_res$ipp), "\\\\")[[1]])
+  iop <- as.numeric(strsplit(rawToChar(dcm_res$iop), "\\\\")[[1]])
+  pixsp <- as.numeric(strsplit(rawToChar(dcm_res$pixsp), "\\\\")[[1]])
+  slice_t <- as.numeric(rawToChar(dcm_res$slice_t))
+  rows <- readBin(dcm_res$rows, "integer", size = 2, signed = FALSE)
+  cols <- readBin(dcm_res$cols, "integer", size = 2, signed = FALSE)
+  slices <- readBin(dcm_res$slices, "integer", size = 2, signed = FALSE)
+  N <- readBin(dcm_res$Npt, "integer")
+  te <- as.numeric(rawToChar(dcm_res$te)) / 1e3
+  tr <- as.numeric(rawToChar(dcm_res$tr)) / 1e3
+  seq_n <- trimws(rawToChar(dcm_res$seq_n))
+  
+  row_ori <- iop[1:3]
+  col_ori <- iop[4:6]
+  sli_vec <- crossprod_3d(row_ori, col_ori)
+  pos_vec <- ipp
+  
+  fids <- rows * cols * slices
+  
+  if (N * 2 * fids * 4 != length(dcm_res$data)) {
+    stop("Unexpected number of data points.")
+  }
+  
+  # read the fid points 
+  raw_vec <- readBin(dcm_res$data, "double", length(dcm_res$data) / 4, size = 4,
+                     endian = "little")
+  
+  data <- raw_vec[c(TRUE, FALSE)] + 1i * raw_vec[c(FALSE, TRUE)]
+  
+  dim(data) <- c(N, rows, cols, slices, 1, 1, 1)
+  data <- aperm(data, c(7, 2, 3, 4, 5, 6, 1))
+  
+  res <- c(NA, pixsp[2], pixsp[1], slice_t, 1, NA, 1 / fs)
+  ref <- def_ref()
+  
+  # TODO determine from the data
+  nuc <- def_nuc()
+  
+  # freq domain vector
+  freq_domain <- rep(FALSE, 7)
+  
+  pos_vec_affine <- pos_vec
+  
+  affine <- cbind(c(row_ori * res[2], 0),
+                  c(col_ori * res[3], 0),
+                  c(sli_vec * res[4], 0),
+                  c(pos_vec_affine, 1))
+  affine[1:2,] <- -affine[1:2,]
+  
+  meta <- list(EchoTime = te,
+               RepetitionTime = tr,
+               Manufacturer = "UIH",
+               SequenceName = seq_n)
+  
+  mrs_data <- mrs_data(data = data, ft = ft, resolution = res, ref = ref,
+                       nuc = nuc, freq_domain = freq_domain, affine = affine,
+                       meta = meta, extra = extra)
+  
+  return(mrs_data)
 }
 
 read_siemens_dicom <- function(fraw, extra, verbose) {
@@ -54,7 +141,7 @@ read_siemens_dicom <- function(fraw, extra, verbose) {
                te      = "0018,9082")
 
   # read em out
-  dcm_res  <- dicom_reader(fraw, tags)
+  dcm_res <- dicom_reader(fraw, tags)
   
   # sort out the right datatypes
   fs  <- readBin(dcm_res$fs, "double")